Add support for verifying a previously written file.
Add support for verifying a previously written file.  A verification header
is written to the beginning of the test file, or to a separate file, and stores
the parameters used for writing the file.  As part of this change, the random
number are generated differently, using a pseudorandom generator rather than
os.urandom().  For performance, a single 'large' buffer of pseudorandom data
is allocated, and random offsets within that buffer are used to write and
verify the data.