Added Tests to check all included config files
test_config_files.py will check each configuration file in the
configfiles directory. To do this it will run them with book register.py
as well as trying to start the librarian with the created db file. By
default this script will check the configfiles directory and additonal
directories can be added by adding the directory path to the parameters
into the function.