Removed Extra skips on mv corner cases
This commit cleaned up the skips that were around the code.
It removed the extra skips in the mv corner cases which are no longer
needed.
This commit also changed what the rename test is doing to avoid possible
conflicts with other tests. test_mv_overwrite is now marked as an
expected failure.